Added special down, fixed animation/projectile bugs

+++ Src/Entity.h
@@ -5,9 +5,9 @@ Base class for all objects with movement and/or borders
 #ifndef ENTITY_H
 #define ENTITY_H
 
-#include "stdafx.h"
-#include "SDL.h"
-//#include <SDL/SDL.h>
+//#include "stdafx.h"
+//#include "SDL.h"
+#include <SDL/SDL.h>
 
 using namespace std;
 
@@ -44,7 +44,7 @@ class Entity {
 	protected:
 		SDL_Surface *sprite;
 
-		int posX, posY;
+		double posX, posY;
 		int height, width;
 
 		int top, bot, left, right;
